pulse_ai/requirements.txt
streamlit==1.32.0
google-generativeai==0.5.2
firebase-admin==6.4.0
plotly==5.20.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1
streamlit-extras==0.4.2
mediapipe==0.10.11
requests==2.31.0
pytz==2024.1
